requirements.txt
python-dotenv~=1.0.1
langchain-openai
langchain-text-splitters
lxml
bs4
langchain_community
unstructured[html]~=0.14.9
psutil
chromadb
tiktoken~=0.7.0
langchain~=0.2.6